Refactor `CommentHandler` module to separate features
