+++ src/js/Content/Features/Community/CCommunityBase.js
@@ -1,6 +1,8 @@
 import {Context, ContextType} from "../../modulesContent";
 import FHideTrademarks from "../Common/FHideTrademarks";
 import FConfirmDeleteComment from "./FConfirmDeleteComment";
+import FHideSpamComments from "./FHideSpamComments";
+import FFavoriteEmoticons from "./FFavoriteEmoticons";
 
 export class CCommunityBase extends Context {
 
@@ -9,6 +11,8 @@ export class CCommunityBase extends Context {
         features.push(
             FHideTrademarks,
             FConfirmDeleteComment,
+            FHideSpamComments,
+            FFavoriteEmoticons,
         );
 
         super(type, features);
