+++ src/js/Content/Features/Community/ProfileActivity/CProfileActivity.js
@@ -1,7 +1,7 @@
 import {ContextType} from "../../../modulesContent";
 import {CCommunityBase} from "../CCommunityBase";
-import {CommentHandler} from "../../../Modules/Community/CommentHandler";
 import FEarlyAccess from "../../Common/FEarlyAccess";
+import FHideSpamComments from "../FHideSpamComments";
 import FHighlightFriendsActivity from "./FHighlightFriendsActivity";
 import FAchievementLink from "./FAchievementLink";
 import FReplaceCommunityHubLinks from "./FReplaceCommunityHubLinks";
@@ -41,12 +41,10 @@ export class CProfileActivity extends CCommunityBase {
     _registerObserver() {
 
         new MutationObserver(mutations => {
-
-            CommentHandler.hideSpamComments();
-
             for (const {addedNodes} of mutations) {
                 this.triggerCallbacks(addedNodes[0]);
                 FEarlyAccess.show(addedNodes[0].querySelectorAll(".blotter_gamepurchase_logo, .gameLogoHolder_default"));
+                FHideSpamComments.handleAllCommentThreads(addedNodes[0]);
             }
         }).observe(document.querySelector("#blotter_content"), {"childList": true});
     }
