+++ src/js/Content/Features/CommunityPage.js
@@ -1,15 +1,5 @@
 import {Page} from "./Page";
-import {CommentHandler} from "../Modules/Community/CommentHandler";
 
-class CommunityPage extends Page {
-
-    _pageSpecificFeatures() {
-        super._pageSpecificFeatures();
-
-        CommentHandler.hideSpamComments();
-        CommentHandler.addFavoriteEmoticons();
-    }
-
-}
+class CommunityPage extends Page {}
 
 export {CommunityPage};
